[#27] - Configuration doesn't use right logging variables
#27

The configuration for the log file was not correct. We were using the
wrong variable which was set at the top of the file. We removed the
wrong variables and set up the correct variable for the log files

We also added a variable in params which is the directory for the log
file.